+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ### Added
 - acp-authz plugin
 - 'components' registry
+- authnId to Cloudentity AuthnCtx
 
 ## [1.1.0] - 2020-04-21
 ### Added